changes files, remove setter from state
