+++ src/routes/Home/Home.jsx
@@ -6,8 +6,8 @@ import Header from '../../components/blocks/Header/Header';
 import Search from '../../components/blocks/Search/Search';
 import InfiniteScroll from '../../components/blocks/InfiniteScroll/InfiniteScroll';
 
-// declaring constants
-const POKE_URL = 'https://pokeapi.co/api/v2/pokemon/';
+// api calls
+import { getAllPokemons, getPokemonFromApi } from '../../api/api';
 
 // App component
 function App() {
@@ -69,7 +69,7 @@ function App() {
         console.log('pokémon not found in stored cache, making request for the api');
         setSearchMessage('pokémon not found in stored cache, making request for the api');
         setLoading(true);
-        getPokemonFromSearch(search.toLowerCase());
+        fetchPokemonData(search.toLowerCase());
       }
     } else {
       setSearchResult([]);
@@ -78,108 +78,35 @@ function App() {
     }
   };
 
-  // get all the pokemon data
-  const getAllPokemons = async (length) => {
-    // defining an empty array to store the promises
-    const promiseArray = [];
-
-    // populate the array with promises
-    // this time, we make the requests for the api based on the id from the pokemon
-    // the try-catch block is insde the foreach, to verify if the first request is succeed or not
-    // if not, throw an error and stops the loop, passing the message to the statee
-    for (let i = length; i < length + 20; i++) {
-      try {
-        // push into the state array
-        promiseArray.push((await fetch(POKE_URL + `${i}`)).json());
-        // error caching
-      } catch (error) {
-        setMessage(error.message);
-        break;
-      }
-    }
-
-    // wait for the promise array to resolve
-    const allPokemonData = await Promise.all(promiseArray);
-
-    // return the data from each pokemon
-    return allPokemonData.map((pokemon) => {
-      return {
-        name: pokemon.name,
-        sprite: pokemon.sprites.front_default,
-        artwork: pokemon.sprites.other['official-artwork'].front_default,
-        stats: pokemon.stats,
-      };
-    });
-  };
-
-  // get the pokemon from the search
-  const getPokemonFromApi = async (pokemonName) => {
-    // try-catch block
-    try {
-      // make the request for the api by pokemon name
-      let promise = await fetch(POKE_URL + pokemonName);
-
-      // waiting the promise to be fullfilled
-      let result = await promise.json();
-
-      // printing the result
-      console.log(result);
-
-      let resultArray = [
-        {
-          name: result.name,
-          sprite: result.sprites.front_default,
-          artwork: result.sprites.other['official-artwork'].front_default,
-          stats: result.stats,
-        },
-      ];
-
-      console.log('Array is:', resultArray);
-
-      return resultArray;
-      // handling errors
-    } catch (error) {
-      // error message
-      console.log(error.message);
-      throw error;
-    }
-  };
-
-  const getPokemonFromSearch = (pokemonName) => {
-    if (pokemonName) {
-      // fetching the data and updating the states
-      const fetchData = async () => {
-        try {
-          const response = await getPokemonFromApi(pokemonName);
-          setSearchResult(response);
-        } catch (err) {
-          console.log(err);
-          setSearchResult([]);
-          setSearchMessage('Pokémon not found!');
-          throw err;
-        }
-      };
-
-      fetchData();
-    }
-  };
-
   // effects
   // using to get the pokemon data
   // useEffect(getPokemonsWhileScroll, []);
-  useEffect(getPokemonFromSearch, []);
+  // useEffect(getPokemonFromSearch, []);
 
   // fetching the data and updating the states
-  const fetchData = async () => {
-    setLoading(true);
-    setMessage('loading');
-    const response = await getAllPokemons(1);
-    setAllPokemons(response);
-    setLoading(false);
+  const fetchPokemonData = async (pokemonName) => {
+    if (pokemonName) {
+      try {
+        const response = await getPokemonFromApi(pokemonName);
+        setSearchResult(response);
+      } catch (err) {
+        console.log(err);
+        setSearchResult([]);
+        setSearchMessage('Pokémon not found!');
+        throw err;
+      }
+    } else {
+      setLoading(true);
+      setMessage('loading');
+      const response = await getAllPokemons(1);
+      setAllPokemons(response);
+      setLoading(false);
+    }
   };
 
+  // useEffect to fetch the data for all the pokemons
   useEffect(() => {
-    fetchData();
+    fetchPokemonData();
   }, []);
 
   // useEffect to handle the scroll event for the user
